chore: added gitignore file-removed unused deps-removed egg dir and test file
