--- vivarium_tellurium.egg-info/PKG-INFO
@@ -1,56 +1,0 @@
-Metadata-Version: 2.1
-Name: vivarium-tellurium
-Version: 0.0.1
-Home-page: https://github.com/vivarium-collective/vivarium-tellurium
-Author: Alex Patrie, Eran Agmon
-Author-email: 
-Description-Content-Type: text/markdown
-
-# Template Vivarium Project
-
-Visit [the Vivarium Core
-documentation](https://vivarium-core.readthedocs.io/) to learn how to
-use the core Vivarium engine to create computational biology models.
-Check out the
-[getting started](https://vivarium-core.readthedocs.io/en/latest/getting_started.html)
-guide of the documentation. 
-
-## Using this Template
-This repository contains a skeleton project you can build on to create
-your own models. Go to the 
-[github page](https://github.com/vivarium-collective/vivarium-template) 
-and select `use this template`.
-
-What follows is a template `README.md`.
-
-<!--
-TODO: Rename the `template` folder to something that makes sense for
-your project.
--->
-
-------------------------------------------------------------------------
-
-# Project Name
-
-<!--
-TODO: Fill in your own project name
--->
-
-Introduction to your project.
-
-<!--
-TODO: Fill in your own introduction.
--->
-
-## Installation
-
-Explain how users can get up and running.
-
-<!--
-TODO: Fill in your own installation instructions. This might be as
-simple as:
-
-```
-pip install <your package name>
-```
--->
