--- vivarium_tellurium.egg-info/SOURCES.txt
@@ -1,14 +1,0 @@
-README.md
-setup.py
-vivarium_tellurium/__init__.py
-vivarium_tellurium.egg-info/PKG-INFO
-vivarium_tellurium.egg-info/SOURCES.txt
-vivarium_tellurium.egg-info/dependency_links.txt
-vivarium_tellurium.egg-info/requires.txt
-vivarium_tellurium.egg-info/top_level.txt
-vivarium_tellurium/composites/__init__.py
-vivarium_tellurium/composites/injected_glc_phosphorylation.py
-vivarium_tellurium/experiments/__init__.py
-vivarium_tellurium/experiments/test_tellurium_process.py
-vivarium_tellurium/processes/__init__.py
-vivarium_tellurium/processes/tellurium_process.py
